Moved installed_sensors.py functions into other appropriate areas Fixed debug log setting not writing properly Refactoring
